New and greatly improved search
Breaks down the level into 5x5 panels, and prioritizes walls (and to a
lesser extent, corridors) that face a empty panel.  This hurts CPU
performance of the search rather noticably; if this proves to be a real
problem, panel occupation status can be cached easily.  Searching is
not common behavior anyways.